Add logger (NLog) and fix bug with controllers #152
