+++ Cooper.Controllers/AuthController.cs
