+++ Cooper.Controllers/Cooper.Controllers.csproj
@@ -6,14 +6,11 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Cooper.Configuration\Cooper.Configuration.csproj" />
     <ProjectReference Include="..\Cooper.Controllers.ViewModels\Cooper.Controllers.ViewModels.csproj" />
     <ProjectReference Include="..\Cooper.Extensions\Cooper.Extensions.csproj" />
-    <ProjectReference Include="..\Cooper.Models\Cooper.Models.csproj" />
     <ProjectReference Include="..\Cooper.ORM\Cooper.ORM.csproj" />
     <ProjectReference Include="..\Cooper.Repositories\Cooper.Repositories.csproj" />
     <ProjectReference Include="..\Cooper.Services.Interfaces\Cooper.Services.Interfaces.csproj" />
