--- Cooper.Logging/Cooper.Logging.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NLog.Extensions.Logging" Version="1.5.2" />
-  </ItemGroup>
-
-</Project>
