+++ Cooper/Cooper.csproj
@@ -15,8 +15,10 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
+    <PackageReference Include="NLog.Web.AspNetCore" Version="4.8.4" />
     <PackageReference Include="Oracle.ManagedDataAccess.Core" Version="2.18.3" />
   </ItemGroup>
 
@@ -34,11 +36,15 @@
   <ItemGroup>
     <ProjectReference Include="..\Cooper.Configuration\Cooper.Configuration.csproj" />
     <ProjectReference Include="..\Cooper.Controllers.ViewModels\Cooper.Controllers.ViewModels.csproj" />
-    <ProjectReference Include="..\Cooper.Controller\Cooper.Controller.csproj" />
-    <ProjectReference Include="..\Cooper.Logging\Cooper.Logging.csproj" />
+    <ProjectReference Include="..\Cooper.Controllers\Cooper.Controllers.csproj" />
+    <ProjectReference Include="..\Cooper.DAO\Cooper.DAO.csproj" />
+    <ProjectReference Include="..\Cooper.Extensions\Cooper.Extensions.csproj" />
     <ProjectReference Include="..\Cooper.Models\Cooper.Models.csproj" />
     <ProjectReference Include="..\Cooper.ORM\Cooper.ORM.csproj" />
     <ProjectReference Include="..\Cooper.Repositories\Cooper.Repositories.csproj" />
+    <ProjectReference Include="..\Cooper.Services.Interfaces\Cooper.Services.Interfaces.csproj" />
+    <ProjectReference Include="..\Cooper.Services\Cooper.Services.csproj" />
+    <ProjectReference Include="..\Cooper.SignalR\Cooper.SignalR.csproj" />
   </ItemGroup>
 
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
