+++ Cooper/Cooper.sln
@@ -15,10 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.DAO", "..\Cooper.DAO
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Controllers.ViewModels", "..\Cooper.Controllers.ViewModels\Cooper.Controllers.ViewModels.csproj", "{7C72AAD9-5E34-4DCB-8A3E-35C612E3BC27}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Logging", "..\Cooper.Logging\Cooper.Logging.csproj", "{795A5807-BB0E-45A7-8062-87B70CD402DC}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Controller", "..\Cooper.Controller\Cooper.Controller.csproj", "{7E7F3108-559E-4D6B-B667-52B42565BBB9}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Extensions", "..\Cooper.Extensions\Cooper.Extensions.csproj", "{2E8E34C9-0EF7-4CD2-BB61-C1AD7CDD996A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.SignalR", "..\Cooper.SignalR\Cooper.SignalR.csproj", "{0FE14F15-5E01-4455-922B-F20A7959F841}"
@@ -27,9 +23,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Services.Interfaces"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Services", "..\Cooper.Services\Cooper.Services.csproj", "{9EC22A77-03EF-4303-A47A-ACCA019E9AC4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cooper.Repositories", "..\Cooper.Repositories\Cooper.Repositories.csproj", "{611C57C5-F789-475B-99F0-45DE3F736AB5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Repositories", "..\Cooper.Repositories\Cooper.Repositories.csproj", "{611C57C5-F789-475B-99F0-45DE3F736AB5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cooper.Tests", "..\Cooper.Tests\Cooper.Tests.csproj", "{8E96DF7D-B633-48F6-B4E1-B673E95341DA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cooper.Tests", "..\Cooper.Tests\Cooper.Tests.csproj", "{8E96DF7D-B633-48F6-B4E1-B673E95341DA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cooper.Controllers", "..\Cooper.Controllers\Cooper.Controllers.csproj", "{DE7F41A9-BD99-4CF7-88C4-645D97254DB1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -61,14 +59,6 @@ Global
 		{7C72AAD9-5E34-4DCB-8A3E-35C612E3BC27}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C72AAD9-5E34-4DCB-8A3E-35C612E3BC27}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7C72AAD9-5E34-4DCB-8A3E-35C612E3BC27}.Release|Any CPU.Build.0 = Release|Any CPU
-		{795A5807-BB0E-45A7-8062-87B70CD402DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{795A5807-BB0E-45A7-8062-87B70CD402DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{795A5807-BB0E-45A7-8062-87B70CD402DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{795A5807-BB0E-45A7-8062-87B70CD402DC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7E7F3108-559E-4D6B-B667-52B42565BBB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7E7F3108-559E-4D6B-B667-52B42565BBB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7E7F3108-559E-4D6B-B667-52B42565BBB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7E7F3108-559E-4D6B-B667-52B42565BBB9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2E8E34C9-0EF7-4CD2-BB61-C1AD7CDD996A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2E8E34C9-0EF7-4CD2-BB61-C1AD7CDD996A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E8E34C9-0EF7-4CD2-BB61-C1AD7CDD996A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -93,6 +83,10 @@ Global
 		{8E96DF7D-B633-48F6-B4E1-B673E95341DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8E96DF7D-B633-48F6-B4E1-B673E95341DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8E96DF7D-B633-48F6-B4E1-B673E95341DA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE7F41A9-BD99-4CF7-88C4-645D97254DB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE7F41A9-BD99-4CF7-88C4-645D97254DB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE7F41A9-BD99-4CF7-88C4-645D97254DB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE7F41A9-BD99-4CF7-88C4-645D97254DB1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
