+++ Cooper/Startup.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using System.Reflection;
 
 [assembly: ApiController]
 namespace Cooper
@@ -14,7 +16,7 @@ namespace Cooper
     {
         public Startup(IConfiguration configuration)
         {
-            Configuration = configuration;
+            Configuration = configuration;  
         }
 
         public IConfiguration Configuration { get; }
@@ -30,7 +32,9 @@ namespace Cooper
             }));
 
             services.AddSignalR();
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            var controllersAssembly = Assembly.Load("Cooper.Controllers");
+            services.AddMvc().AddApplicationPart(controllersAssembly).AddControllersAsServices();
             services.AddSingleton<ICommonChatRepository, CommonChatRepository>();
 
             // In production, the Angular files will be served from this directory
@@ -47,6 +51,8 @@ namespace Cooper
             services.AddUserConnectionService();
             services.AddTokenCleanerService();
             services.AddSocialAuthService();
+
+
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
